shifted genesis config to casper_types
