+++ execution_engine_testing/test_support/src/auction.rs
@@ -5,10 +5,7 @@ use rand::Rng;
 use tempfile::TempDir;
 
 use casper_execution_engine::engine_state::{
-    engine_config::{DEFAULT_FEE_HANDLING, DEFAULT_REFUND_HANDLING},
-    genesis::ExecConfigBuilder,
-    run_genesis_request::RunGenesisRequest,
-    EngineState, ExecuteRequest,
+    run_genesis_request::RunGenesisRequest, EngineState, ExecuteRequest,
 };
 use casper_storage::global_state::{
     state::{CommitProvider, StateProvider},
@@ -19,17 +16,17 @@ use casper_types::{
     bytesrepr::{self},
     runtime_args,
     system::auction,
-    ChainspecRegistry, Digest, GenesisAccount, GenesisValidator, Key, Motes, ProtocolVersion,
-    PublicKey, SecretKey, StoredValue, U512,
+    ChainspecRegistry, Digest, GenesisAccount, GenesisConfigBuilder, GenesisValidator, Key, Motes,
+    ProtocolVersion, PublicKey, SecretKey, StoredValue, DEFAULT_REFUND_HANDLING, U512,
 };
 
 use crate::{
     transfer, DeployItemBuilder, ExecuteRequestBuilder, LmdbWasmTestBuilder, StepRequestBuilder,
     DEFAULT_ACCOUNT_ADDR, DEFAULT_ACCOUNT_INITIAL_BALANCE, DEFAULT_ACCOUNT_PUBLIC_KEY,
-    DEFAULT_AUCTION_DELAY, DEFAULT_GENESIS_CONFIG_HASH, DEFAULT_GENESIS_TIMESTAMP_MILLIS,
-    DEFAULT_LOCKED_FUNDS_PERIOD_MILLIS, DEFAULT_PROPOSER_PUBLIC_KEY, DEFAULT_PROTOCOL_VERSION,
-    DEFAULT_ROUND_SEIGNIORAGE_RATE, DEFAULT_SYSTEM_CONFIG, DEFAULT_UNBONDING_DELAY,
-    DEFAULT_WASM_CONFIG, SYSTEM_ADDR,
+    DEFAULT_AUCTION_DELAY, DEFAULT_FEE_HANDLING, DEFAULT_GENESIS_CONFIG_HASH,
+    DEFAULT_GENESIS_TIMESTAMP_MILLIS, DEFAULT_LOCKED_FUNDS_PERIOD_MILLIS,
+    DEFAULT_PROPOSER_PUBLIC_KEY, DEFAULT_PROTOCOL_VERSION, DEFAULT_ROUND_SEIGNIORAGE_RATE,
+    DEFAULT_SYSTEM_CONFIG, DEFAULT_UNBONDING_DELAY, DEFAULT_WASM_CONFIG, SYSTEM_ADDR,
 };
 
 const ARG_AMOUNT: &str = "amount";
@@ -314,7 +311,7 @@ fn create_run_genesis_request(
     validator_slots: u32,
     genesis_accounts: Vec<GenesisAccount>,
 ) -> RunGenesisRequest {
-    let exec_config = ExecConfigBuilder::default()
+    let exec_config = GenesisConfigBuilder::default()
         .with_accounts(genesis_accounts)
         .with_wasm_config(*DEFAULT_WASM_CONFIG)
         .with_system_config(*DEFAULT_SYSTEM_CONFIG)
