+++ execution_engine_testing/test_support/src/utils.rs
@@ -9,13 +9,13 @@ use std::{
 use once_cell::sync::Lazy;
 
 use casper_execution_engine::engine_state::{
-    engine_config::{DEFAULT_FEE_HANDLING, DEFAULT_REFUND_HANDLING},
-    execution_result::ExecutionResult,
-    genesis::{ExecConfig, ExecConfigBuilder, GenesisConfig},
-    run_genesis_request::RunGenesisRequest,
-    Error,
+    execution_result::ExecutionResult, genesis::GenesisRequest,
+    run_genesis_request::RunGenesisRequest, Error,
 };
-use casper_types::{Gas, GenesisAccount};
+use casper_types::{
+    Gas, GenesisAccount, GenesisConfig, GenesisConfigBuilder, DEFAULT_FEE_HANDLING,
+    DEFAULT_REFUND_HANDLING,
+};
 
 use super::{DEFAULT_ROUND_SEIGNIORAGE_RATE, DEFAULT_SYSTEM_CONFIG, DEFAULT_UNBONDING_DELAY};
 use crate::{
@@ -131,7 +131,7 @@ pub fn read_wasm_file_bytes<T: AsRef<Path>>(contract_file: T) -> Vec<u8> {
 }
 
 /// Returns an [`ExecConfig`].
-pub fn create_exec_config(accounts: Vec<GenesisAccount>) -> ExecConfig {
+pub fn create_exec_config(accounts: Vec<GenesisAccount>) -> GenesisConfig {
     let wasm_config = *DEFAULT_WASM_CONFIG;
     let system_config = *DEFAULT_SYSTEM_CONFIG;
     let validator_slots = DEFAULT_VALIDATOR_SLOTS;
@@ -143,7 +143,7 @@ pub fn create_exec_config(accounts: Vec<GenesisAccount>) -> ExecConfig {
     let refund_handling = DEFAULT_REFUND_HANDLING;
     let fee_handling = DEFAULT_FEE_HANDLING;
 
-    ExecConfigBuilder::default()
+    GenesisConfigBuilder::default()
         .with_accounts(accounts)
         .with_wasm_config(wasm_config)
         .with_system_config(system_config)
@@ -158,14 +158,14 @@ pub fn create_exec_config(accounts: Vec<GenesisAccount>) -> ExecConfig {
         .build()
 }
 
-/// Returns a [`GenesisConfig`].
-pub fn create_genesis_config(accounts: Vec<GenesisAccount>) -> GenesisConfig {
+/// Returns a [`GenesisRequest`].
+pub fn create_genesis_config(accounts: Vec<GenesisAccount>) -> GenesisRequest {
     let name = DEFAULT_CHAIN_NAME.to_string();
     let timestamp = DEFAULT_GENESIS_TIMESTAMP_MILLIS;
     let protocol_version = *DEFAULT_PROTOCOL_VERSION;
     let exec_config = create_exec_config(accounts);
 
-    GenesisConfig::new(name, timestamp, protocol_version, exec_config)
+    GenesisRequest::new(name, timestamp, protocol_version, exec_config)
 }
 
 /// Returns a [`RunGenesisRequest`].
